+++ cmd/livesim2/app/livempd.go
@@ -204,6 +204,10 @@ func LiveMPD(a *asset, mpdName string, cfg *ResponseConfig, drmCfg *drm.DrmConfi
 		addSGAIReplaceEvents(mpd, period, cfg, endTimeMS)
 	}
 
+	if cfg.SVTA != nil {
+		addSVTAAdCreativeEvents(mpd, period, cfg, endTimeMS)
+	}
+
 	if cfg.Steer != nil {
 		addContentSteering(mpd, period, cfg)
 	}